🚚 : split shared components in subfolders
