+++ src/main/client/app/shared/components/cli-badge/cli-badge.vue
