+++ src/main/client/app/shared/components/index.js
@@ -3,7 +3,7 @@ export { default as AppPageForbidden } from '@/shared/components/error-pages/pag
 export { default as AppTopBar } from '@/shared/components/topbar/top-bar.vue';
 export { default as AppSideBar } from '@/shared/components/sidebar/side-bar.vue';
 export { default as AppBreadcrumb } from '@/shared/components/breadcrumb/breadcrumb.vue';
-export { default as AppCliBadge } from '@/shared/components/cli-badge.vue';
-export { default as AppFormTypeahead } from '@/shared/components/form-typeahead.vue';
-export { default as AppMarkdown } from '@/shared/components/markdown.vue';
+export { default as AppCliBadge } from '@/shared/components/cli-badge/cli-badge.vue';
+export { default as AppFormTypeahead } from '@/shared/components/typeahead/form-typeahead.vue';
+export { default as AppMarkdown } from '@/shared/components/markdown/markdown.vue';
 export { default as AppConsole } from '@/shared/components/console/console.vue';
