+++ src/main/client/app/shared/components/markdown/markdown.vue
